client: Remove ImageCreate in favor of ImagePull/ImageImport
