--- client/image_create_test.go
@@ -1,65 +1,0 @@
-package client
-
-import (
-	"context"
-	"fmt"
-	"io"
-	"net/http"
-	"testing"
-
-	cerrdefs "github.com/containerd/errdefs"
-	"github.com/moby/moby/api/types/registry"
-	"gotest.tools/v3/assert"
-	is "gotest.tools/v3/assert/cmp"
-)
-
-func TestImageCreateError(t *testing.T) {
-	client, err := New(WithMockClient(errorMock(http.StatusInternalServerError, "Server error")))
-	assert.NilError(t, err)
-	_, err = client.ImageCreate(context.Background(), "reference", ImageCreateOptions{})
-	assert.Check(t, is.ErrorType(err, cerrdefs.IsInternal))
-}
-
-func TestImageCreate(t *testing.T) {
-	const (
-		expectedURL          = "/images/create"
-		expectedImage        = "docker.io/test/my_image"
-		expectedTag          = "sha256:ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
-		specifiedReference   = "test/my_image:latest@" + expectedTag
-		expectedRegistryAuth = "eyJodHRwczovL2luZGV4LmRvY2tlci5pby92MS8iOnsiYXV0aCI6ImRHOTBid289IiwiZW1haWwiOiJqb2huQGRvZS5jb20ifX0="
-	)
-
-	client, err := New(WithMockClient(func(req *http.Request) (*http.Response, error) {
-		if err := assertRequest(req, http.MethodPost, expectedURL); err != nil {
-			return nil, err
-		}
-		registryAuth := req.Header.Get(registry.AuthHeader)
-		if registryAuth != expectedRegistryAuth {
-			return nil, fmt.Errorf("%s header not properly set in the request. Expected '%s', got %s", registry.AuthHeader, expectedRegistryAuth, registryAuth)
-		}
-
-		query := req.URL.Query()
-		fromImage := query.Get("fromImage")
-		if fromImage != expectedImage {
-			return nil, fmt.Errorf("fromImage not set in URL query properly. Expected '%s', got %s", expectedImage, fromImage)
-		}
-
-		tag := query.Get("tag")
-		if tag != expectedTag {
-			return nil, fmt.Errorf("tag not set in URL query properly. Expected '%s', got %s", expectedTag, tag)
-		}
-
-		return mockResponse(http.StatusOK, nil, "body")(req)
-	}))
-	assert.NilError(t, err)
-
-	createResult, err := client.ImageCreate(context.Background(), specifiedReference, ImageCreateOptions{
-		RegistryAuth: expectedRegistryAuth,
-	})
-	assert.NilError(t, err)
-	response, err := io.ReadAll(createResult.Body)
-	assert.NilError(t, err)
-	err = createResult.Body.Close()
-	assert.NilError(t, err)
-	assert.Check(t, is.Equal(string(response), "body"))
-}
